app/src/test/java/com/ml/spaceflightapp/viewmodel/FlightViewModelTest.kt
package com.ml.spaceflightapp.viewmodel

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.lifecycle.Observer
import com.ml.chucknorrisapp.util.CoroutineTestRule
import com.ml.chucknorrisapp.util.LifeCycleTestOwner
import com.ml.spaceflightapp.repository.FlightRepository
import com.ml.spaceflightapp.util.LoadingState
import kotlinx.coroutines.test.runBlockingTest
import org.junit.After
import org.junit.Test

import org.junit.Assert.*
import org.junit.Before
import org.junit.Rule
import org.junit.rules.TestRule
import org.mockito.Mockito

class FlightViewModelTest {

    @get:Rule
    val coroutineTestRule = CoroutineTestRule()

    @get:Rule
    val instantTaskExecutorRule: TestRule = InstantTaskExecutorRule()

    private lateinit var loadingStateObserver: Observer<LoadingState>

    private lateinit var flightRepository: FlightRepository

    private lateinit var flightViewModel: FlightViewModel

    private lateinit var lifeCycleTestOwner: LifeCycleTestOwner


    @Before
    fun setUp() {
        lifeCycleTestOwner = LifeCycleTestOwner()
        lifeCycleTestOwner.onCreate()

        flightRepository = Mockito.mock(FlightRepository::class.java)

        flightViewModel = FlightViewModel(flightRepository)

        loadingStateObserver = Mockito.mock(Observer::class.java) as Observer<LoadingState>
        flightViewModel.loadingState.observe(lifeCycleTestOwner, loadingStateObserver)
    }

    @Test
    fun whenGetFlightData_thenGetFlightsCalled(){
        coroutineTestRule.testDispatcher.runBlockingTest {

            // Given
            lifeCycleTestOwner.onResume()

            Mockito.`when`(flightRepository.getFlights()).thenReturn(null)

            // When
            flightViewModel.getFlightData()

            // Then
            Mockito.verify(flightRepository, Mockito.times(2)).getFlights()
            Mockito.verify(loadingStateObserver).onChanged(LoadingState.LOADING)
        }
    }

    @After
    fun tearDown(){
        lifeCycleTestOwner.onDestroy()
    }
}